+++ reana-ui/src/pages/workflowDetails/WorkflowDetails.js
@@ -2,19 +2,24 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020 CERN.
+  Copyright (C) 2020, 2022 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
 */
 
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
 import { Container, Dimmer, Loader, Tab } from "semantic-ui-react";
 
-import { fetchWorkflow } from "~/actions";
-import { getWorkflow, loadingWorkflows, isWorkflowsFetched } from "~/selectors";
+import { fetchWorkflow, fetchWorkflowLogs } from "~/actions";
+import {
+  getWorkflow,
+  loadingWorkflows,
+  isWorkflowsFetched,
+  getConfig,
+} from "~/selectors";
 import BasePage from "../BasePage";
 import { Notification, WorkflowDeleteModal } from "~/components";
 import {
@@ -32,6 +37,8 @@ export default function WorkflowDetailsPage() {
   );
 }
 
+const FINISHED_STATUSES = ["finished", "failed", "stopped", "deleted"];
+
 function WorkflowDetails() {
   const { id: workflowId } = useParams();
 
@@ -39,12 +46,36 @@ function WorkflowDetails() {
   const workflow = useSelector(getWorkflow(workflowId));
   const loading = useSelector(loadingWorkflows);
   const workflowsFetched = useSelector(isWorkflowsFetched);
+  const { pollingSecs } = useSelector(getConfig);
+  const interval = useRef(null);
 
   // TODO: add `workflowRefresh` dependency to trigger useEffect
   // once user opens, closes Interactive session, deletes workflow
   useEffect(() => {
     dispatch(fetchWorkflow(workflowId));
-  }, [dispatch, workflowId]);
+    if (!interval.current && pollingSecs) {
+      interval.current = setInterval(() => {
+        dispatch(
+          fetchWorkflow(workflowId, { refetch: true, showLoader: false })
+        );
+        dispatch(
+          fetchWorkflowLogs(workflowId, { refetch: true, showLoader: false })
+        );
+      }, pollingSecs * 1000);
+    }
+    return cleanPolling;
+  }, [dispatch, workflowId, pollingSecs]);
+
+  useEffect(() => {
+    if (workflow && FINISHED_STATUSES.includes(workflow.status)) {
+      cleanPolling();
+    }
+  }, [workflow]);
+
+  const cleanPolling = () => {
+    clearInterval(interval.current);
+    interval.current = null;
+  };
 
   if (!workflowsFetched || loading) {
     return (