+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.9.0 (UNRELEASED)
 - Adds a label link to the workflow launcher URL remote origin on the workflow details page.
 - Adds Launch on REANA page allowing the submission of workflows via badge-clicking.
 - Adds 404 Not Found error page.
+- Adds periodic refresh of the workflow details page.
 - Fixes redirection chain for non-signed-in CERN SSO users to access the desired target page after sign-in.
 - Fixes ``fetchWorkflow`` action to fetch a specific workflow instead of the entire user workflow list.